requirements.txt
pandas
dashscope
python-dotenv
matplotlib
seaborn
tabulate